Setup eventUpdate mutation.
Fundamentally, I'd assert that all actions associated with the execution of a mutation should appear in the resolver ("source"), but I'm about to embark on an experiment with nesting those actions in resolvers on the input types. Venue creation will be handled in the VenueInput type. Hold on to your hats! What could possibly go wrong?! ⚗️🔬